mem0-mcp-clean/requirements.txt
# MCP-Mem0 Python依赖
# 核心MCP框架
mcp>=1.12.2

# Web框架
fastapi>=0.104.0
uvicorn>=0.24.0
starlette>=0.27.0

# HTTP客户端
httpx>=0.25.0
httpx-sse>=0.4.0

# 数据处理
pydantic>=2.4.0
pydantic-settings>=2.0.0

# 异步支持
anyio>=4.0.0

# JSON处理
jsonschema>=4.19.0

# 环境变量
python-dotenv>=1.0.0

# Mem0 AI (移除，只使用本地HTTP模式)
# mem0ai>=0.1.0

# 其他工具
python-multipart>=0.0.6
sse-starlette>=1.6.0
